+++ snn-anatomy/src/main/java/com/uwe_hennig/snn/anatomy/neuron/WeightPersistence.java
@@ -18,12 +18,13 @@ import java.nio.file.Path;
 
 /**
  * WeightPersistence
+ *
  * @author Uwe Hennig
  */
 public final class WeightPersistence {
     static void save(WeightModel model, Path file) throws IOException {
-        try (FileChannel channel = FileChannel.open(file, CREATE, READ, WRITE,TRUNCATE_EXISTING)) {
-            long size = model.segment.byteSize();
+        try (FileChannel channel = FileChannel.open(file, CREATE, READ, WRITE, TRUNCATE_EXISTING)) {
+            int size = (int) model.segment.byteSize();
 
             try (Arena tempArena = Arena.ofConfined()) {
                 MemorySegment mappedFile = channel.map(FileChannel.MapMode.READ_WRITE, 0, size, tempArena);
@@ -35,8 +36,8 @@ public final class WeightPersistence {
 
     static WeightModel load(Path file) throws IOException {
         try (FileChannel channel = FileChannel.open(file, READ)) {
-            long size = channel.size();
-            long capacity = size / WeightModel.LAYOUT.byteSize();
+            int size = (int) channel.size();
+            int capacity = size / (int) WeightModel.LAYOUT.byteSize();
             WeightModel model = new WeightModel(capacity);
 
             try (Arena tempArena = Arena.ofConfined()) {
